+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.6.0...main)
 * _Contributing to this repo? Add info about your change here to be included in the next release_
 
+__Improvements__
+- Make relative query take a QueryConstraint as an argument. Add more documentation  ([#129](https://github.com/parse-community/Parse-Swift/pull/129)), thanks to [Corey Baker](https://github.com/cbaker6).
+
 ### 1.6.0
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.5.1...1.6.0)